tools/market-research/.skills/nemotron-instagram-persona/requirements.txt
datasets>=2.14.0
requests>=2.31.0
python-dotenv>=1.0.0
